bump: auth api revamp
