+++ components/categories/CategoryHeader.tsx
@@ -1,8 +1,7 @@
 "use client";
 import React from "react";
-import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
 import { Urbanist } from "next/font/google";
-import { useRouter } from "next/navigation";
+import { useAuth } from "@/lib/hooks/useAuth";
 
 const bodyText = Urbanist({
   weight: "400",
@@ -19,15 +18,10 @@ interface CategoryHeaderProps {
 }
 
 export default function CategoryHeader(props: CategoryHeaderProps) {
-  const router = useRouter();
-
-  // Create a Supabase client configured to use cookies
-  const supabase = createClientComponentClient();
+  const { signOut } = useAuth();
 
   const handleSignOut = async () => {
-    await supabase.auth.signOut();
-    router.push("/login");
-    router.refresh();
+    await signOut();
   };
  
   return (
